Move toArray helpers from AnyList into ListProxy so they are available for all lists.
